feat(plugin): 工具页专用 UI 通用化 + 商店版本标注
- tool_ui registry：插件声明 tool_ui 值 + 前端注册组件即得专用工具卡，
  不再硬编码 tunnel-card（TunnelCard 抽为可复用组件）
- 设置页"外网接入"选项卡删除，工具 Tab 为唯一入口
- 商店安装/更新：needs_core_update 时二次确认"需 DiceFrame ≥ X"
- 商店包体上限放宽（20MB→100MB）支持带二进制大插件
- PluginMarketplaceItem/PluginToolDescriptor 补 tool_ui 字段
