+++ frontend-v2/src/features/plugins/PluginSettings.vue
@@ -22,6 +22,7 @@ import { usePluginTools } from './usePluginTools'
 import { usePluginExport } from './usePluginExport'
 import { usePluginTypes } from './usePluginTypes'
 import { usePluginUninstallCleanup } from './usePluginUninstallCleanup'
+import TunnelCard from './TunnelCard.vue'
 
 const { t } = useLocale()
 const { pluginThemes, pluginThemeId, loadPluginThemes, applyPluginTheme, clearPluginTheme } = useTheme()
@@ -105,6 +106,20 @@ const themeOptions = computed(() => pluginThemes.value.map(theme => ({
 function permissionDescription(p: PluginInfo, permission: string): string {
   return p.permission_details?.find(item => item.id === permission)?.description || permission
 }
+
+// 工具页专用 UI registry：tool_ui 值 -> 渲染组件。未来进程插件声明新的 tool_ui
+// 值并在此注册组件即可获得专用卡，无需改工具页分发逻辑。
+const toolUiRegistry: Record<string, Component> = {
+  'tunnel-card': TunnelCard,
+}
+// 声明了专用工具 UI 且 registry 有对应组件的已装插件：工具 Tab 渲染专用卡而非裸工具表单。
+const toolUiPlugins = computed(() =>
+  plugins.value.filter(p => p.tool_ui && toolUiRegistry[p.tool_ui]),
+)
+// 裸工具网格：过滤掉由专用 UI 卡覆盖的插件的工具（如 cloudflare-tunnel 的 tunnel_*）。
+const plainTools = computed(() =>
+  tools.value.filter(tool => !(tool.tool_ui && toolUiRegistry[tool.tool_ui])),
+)
 
 function selectedThemeDescription(): string {
   const theme = pluginThemes.value.find(item => item.id === pluginThemeId.value)
@@ -419,8 +434,13 @@ onMounted(async () => {
         </NButton>
       </section>
       <NSpin :show="toolsLoading">
-        <div v-if="tools.length" class="tool-grid">
-          <article v-for="tool in tools" :key="toolKey(tool)" class="tool-card">
+        <component
+          v-for="p in toolUiPlugins"
+          :key="p.id"
+          :is="toolUiRegistry[p.tool_ui!]"
+        />
+        <div v-if="plainTools.length" class="tool-grid">
+          <article v-for="tool in plainTools" :key="toolKey(tool)" class="tool-card">
             <div class="tool-heading">
               <div>
                 <h3>{{ tool.title || tool.name }}</h3>
@@ -449,7 +469,7 @@ onMounted(async () => {
             <pre v-if="toolResults[toolKey(tool)]" class="tool-result">{{ toolResults[toolKey(tool)] }}</pre>
           </article>
         </div>
-        <p v-else class="muted">{{ t('noRunningPluginTools') }}</p>
+        <p v-else-if="!toolUiPlugins.length" class="muted">{{ t('noRunningPluginTools') }}</p>
       </NSpin>
     </NTabPane>
