+++ frontend-v2/src/i18n/messages/en.ts
@@ -468,8 +468,6 @@ export const en = {
   publicBaseUrlPlaceholder: 'For example https://trpg.example.com or http://nas.local:18000',
   publicBaseUrlHint: 'Blank: use the current page address ({origin}). Do not include /#/join; the invite path is appended automatically. If deployed under a reverse-proxy subpath, you can use https://example.com/trpg.',
   saveSharingAddress: 'Save Share Address',
-  tunnelTitle: 'External Access',
-  tunnelHelp: 'Generate a public HTTPS address via a Cloudflare quick tunnel with one click — no account or domain needed. Quick tunnels use a random domain that changes on each restart.',
   tunnelEnable: 'Enable',
   tunnelStop: 'Stop',
   tunnelStarting: 'Starting tunnel…',
@@ -704,6 +702,7 @@ export const en = {
   pluginUpdateApprovalRequired: 'Needs review',
   confirmProcessPluginInstall: '“{name}” launches a third-party process with the files and network available to your operating-system account. Install only if you trust the publisher. Continue?',
   confirmProcessPluginUpdate: '“{name}” is a high-privilege process plugin and is never silently updated. Continue only after trusting its latest Release. Update now?',
+  confirmCoreUpgrade: '“{name}” requires DiceFrame {version} or later; your current version may not support it. Install/update anyway?',
   themes: 'Themes',
   pluginThemes: 'Plugin Themes',
   pluginThemesHelp: 'Theme plugins override DiceFrame CSS variables; the selection is stored only in this browser.',
