+++ frontend-v2/src/i18n/messages/zh-CN.ts
@@ -468,8 +468,6 @@ export const zhCN = {
   publicBaseUrlPlaceholder: '例如 https://trpg.example.com 或 http://nas.local:18000',
   publicBaseUrlHint: '留空：使用当前页面地址（{origin}）。不要填写 /#/join，系统会自动拼接邀请路径；如果部署在反向代理子路径，可以填写 https://example.com/trpg。',
   saveSharingAddress: '保存分享地址',
-  tunnelTitle: '外网接入',
-  tunnelHelp: '通过 Cloudflare 快速隧道一键生成公网 HTTPS 地址，免账号免域名。每次启动为随机域名，重启后链接变化。',
   tunnelEnable: '一键启用',
   tunnelStop: '停止',
   tunnelStarting: '隧道启动中…',
@@ -704,6 +702,7 @@ export const zhCN = {
   pluginUpdateApprovalRequired: '待审核',
   confirmProcessPluginInstall: '“{name}”会启动第三方进程，并拥有当前系统用户可访问的文件和网络权限。仅在信任作者时安装。继续吗？',
   confirmProcessPluginUpdate: '“{name}”是高权限进程插件，不会静默更新。确认你信任其最新 Release 后再继续。现在更新吗？',
+  confirmCoreUpgrade: '“{name}”需要 DiceFrame {version} 及以上版本，当前版本较低，可能无法正常工作。仍要安装/更新吗？',
   themes: '主题',
   pluginThemes: '插件主题',
   pluginThemesHelp: '主题插件会覆盖 DiceFrame 的 CSS 变量；选择只保存在当前浏览器。',
